Some static_cast's to make gcc happy.
svn path=/trunk/kdelibs/; revision=44379

+++ kdecore/krootprop.cpp
@@ -342,12 +342,12 @@ QString KRootProp::writeEntry( const QString& rKey, const QFont& rFont )
 
   QString aCharset = "default";
   if( rFont.charSet() != QFont::AnyCharSet )
-      aCharset.setNum( rFont.charSet() );
+      aCharset.setNum( static_cast<int>(rFont.charSet()) );
   
   QTextIStream ts( &aValue );
   ts << rFont.family() << "," << rFont.pointSize() << "," 
-     << rFont.styleHint() << "," << aCharset << "," << rFont.weight() << "," 
-     << nFontBits;
+     << static_cast<int>(rFont.styleHint()) << "," << aCharset << "," << rFont.weight() << "," 
+     << static_cast<int>(nFontBits);
   return writeEntry( rKey, aValue );
 }
